files/scripts/ansible-ini2yaml/requirements.txt
pyyaml
six
ansible
